% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heinzWrapper.R
\name{findHeinzMStTP}
\alias{findHeinzMStTP}
\title{use the heinz algorithm to find a solution to the minimum steiner problem}
\usage{
findHeinzMStTP(
  searchGraph.igraph,
  n.threads = 2,
  induceSubgraph = FALSE,
  verbose = TRUE,
  timeLimit = 600
)
}
\arguments{
\item{searchGraph.igraph}{The igraph representation of the graph that we look to search over with the MStTP solver. Note, this graph should have vertex attributes of 'isTerminal'}

\item{n.threads}{Specifies the maximum number of threads that the heinz binary can use}

\item{induceSubgraph}{Flag to dictate if you would prefer a tree or the induced subgraph of the solution}

\item{verbose}{A flag that specifies if you want information on what the function is doing}

\item{timeLimit}{The maximum number of seconds that the heinz binary should be allowed to run for (default: 600 seconds)}
}
\value{
igraphModule An igraph representation of the solution (with induced edges from the original search graph if requested)
}
\description{
heinz ("heavy induced subgraphs") computes solutions to the minimum Steiner Tree Problem.
}
\references{
An algorithmic framework for the exact solution of the prize-collecting Steiner tree problem. Mathematical Programming, Series B, 105(2-3):427-449, 2006, by I. Ljubic, R. Weiskircher, U. Pferschy, G. Klau, P. Mutzel, and M. Fischetti.

M. T. Dittrich, G. W. Klau, A. Rosenwald, T. Dandekar and T. Mueller (2008) Identifying functional modules in protein-protein interaction networks: an integrated exact approach. (ISMB2008) Bioinformatics 24: 13. i223-i231 Jul.

D. Beisser, G. W. Klau, T. Dandekar, T. Mueller and M. Dittrich (2010) BioNet: an R-package for the Functional Analysis of Biological Networks. Bioinformatics 26:08. 1129-1130 Apr.

\url{http://dimacs11.cs.princeton.edu/contest/challenge-results.pdf}
}
\concept{minimumSteinerTrees}
