% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene42_igraph.R
\docType{data}
\name{gene42_igraph}
\alias{gene42_igraph}
\title{gene42 example from SteinLib}
\format{An igraph object with 335 nodes and 456 edges
\describe{
  \item{name}{Unique identifiers for each vertex - note, not an index despite being character numbers}
  \item{isTerminal}{A boolean flag detailing as to whether a vertex is a fixed node or not}
  \item{edgeWeight}{A belief score between 0 and 1 indicated how strong the evidence is for an interaction}
}}
\source{
\url{http://steinlib.zib.de/showset.php?GENE}
}
\usage{
data(gene42_igraph)
}
\description{
A directed graph representing a Steiner arborescence problem (directed Steiner problems) from a genetic context. This corresponds to gene42 from the steinLib dataset.
}
\details{
Note: it is only advised to use this, rather difficult, Steiner tree problem in conjunction with an advanced MILP solver like CPLEX.
}
\keyword{datasets}
