% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeCentricSteinerForestProblem-class.R
\docType{class}
\name{nodeCentricSteinerForestProblem}
\alias{nodeCentricSteinerForestProblem}
\title{Solve multiple bootstrap Minimum Steiner Tree problems (aka Steiner Forest procedure)}
\format{
R6Class \code{nodeCentricSteinerForestProblem} Construct an object representation of the bootstraped Steiner Tree process (aka Steiner Forest routine)
}
\description{
Given a set of seeds/fixed terminals a Minimum Steiner Tree can be found. One might well be interested in studying the common nodes that
would be included with, say, just 50% of the seed set. This process is known as a 'bootstrap' in statistics and this class looks to
repeatedly sample seeds to produce a consensus set of MStTP solutions. Sub-solutions can also be collected, albeit at an increased burden on the
solver (and therefore dramatically increasing the time).
}
\details{
This class is derived from *subOptimalSteinerProblem* and in turn *nodeCentricSteinerTreeProblem*: all methods available in the superclass are available here. The difference is that after
each acceptable solution is found, the solution is a.) stored in a bootstrap solution pool and b.) used to generate a 'novelty' constraint on future solutions. For each bootstrap run, the solution pool
is flushed and the process re-rerun. In the end, all of the boostrap solutions are in the bootstrap solution pool.
}
\section{methods}{

Alongisde those for *nodeCentricSteinerTreeProblem* and *subOptimalSteinerProblem*
\describe{
   \item{\code{new(network, solverChoice = chooseSolver(), verbose = TRUE, solverTimeLimit = 300, solverTrace = as.integer(verbose), solutionTolerance = 0)}}{Constructor for the nodeCentricSteinerForestProblem class. Note the loss of 'presolveGraph'; the repeated resampling of fixed terminal nodes prevents this.}
   \item{\code{sampleMultipleBootstrapSteinerSolutions(nBootstraps = 5, maxItr = 0, resamplingProbability= 0.5)}}{Run the bootstrap procedure nBootstraps times, each time resampling seeds with pSuccess = resamplingProbability, collecting degenerate or suboptimal solutions for maxItr times.}
   \item{\code{getBootstrapSolutionPoolGraphs(collapseSols = TRUE)}}{Either return a list of solutions within tolerance (collapseSols = FALSE) or pool all solutions together and return a single graph (collapseSols = TRUE, defaults)}
   \item{...}{Other methods are self explanatory and likely uninteresting to a general user}
' }
}

\examples{
library(igraph)

#Prepare a simple seed-based Steiner sampling in a reasonable sized network

fixedTerminalLymphomaGraph = lymphomaGraph
V(fixedTerminalLymphomaGraph)$isTerminal = FALSE
V(fixedTerminalLymphomaGraph)[nodeScore > 0]$isTerminal = TRUE
fixedTerminalLymphomaGraph = delete_vertex_attr(fixedTerminalLymphomaGraph, "nodeScore")


# Example of solving *just* the single-solution Minimum Steiner Tree Problem
MStTPsingle = nodeCentricSteinerTreeProblem$new(fixedTerminalLymphomaGraph)
MStTPsingle$findSingleSteinerSolution()

#Solve multiple bootstrap Steiner Trees (Steiner Forest)
SteinFor = nodeCentricSteinerForestProblem$new(fixedTerminalLymphomaGraph)

#Run two bootstrap routines (resample fixed terminals and solve) and 
#ALSO run the sub-optimal solution searcher thrice
SteinFor$sampleMultipleBootstrapSteinerSolutions(nBootstraps = 2, maxItr = 3)
## Takes around a minute using RGLPK as the solver

}
\references{
Fischetti M, Leitner M, Ljubić I, Luipersbeck M, Monaci M, Resch M, et al. Thinning out Steiner trees: a node-based model for uniform edge costs. Math Program Comput. dimacs11.cs.princeton.edu; 2017;9: 203–229.

Beisser D, Klau GW, Dandekar T, Müller T, Dittrich MT. BioNet: An R-Package for the functional analysis of biological networks. Bioinformatics. 2010;26: 1129–1130.

\url{https://en.wikipedia.org/wiki/Steiner_tree_problem}
}
\seealso{
nodeCentricSteinerTreeProblem

subOptimalSteinerProblem

Other SteinerProblemSolver: 
\code{\link{nodeCentricSteinerTreeProblem}},
\code{\link{subOptimalSteinerProblem}}
}
\concept{SteinerProblemSolver}
\section{Super classes}{
\code{stoneTrees::nodeCentricSteinerTreeProblem} -> \code{stoneTrees::subOptimalSteinerProblem} -> \code{nodeCentricSteinerForestProblem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-nodeCentricSteinerForestProblem-new}{\code{nodeCentricSteinerForestProblem$new()}}
\item \href{#method-nodeCentricSteinerForestProblem-sampleMultipleBootstrapSteinerSolutions}{\code{nodeCentricSteinerForestProblem$sampleMultipleBootstrapSteinerSolutions()}}
\item \href{#method-nodeCentricSteinerForestProblem-getBootstrapSolutionPool}{\code{nodeCentricSteinerForestProblem$getBootstrapSolutionPool()}}
\item \href{#method-nodeCentricSteinerForestProblem-getNconnectivityConstraintsCallsPool}{\code{nodeCentricSteinerForestProblem$getNconnectivityConstraintsCallsPool()}}
\item \href{#method-nodeCentricSteinerForestProblem-getSolutionPool}{\code{nodeCentricSteinerForestProblem$getSolutionPool()}}
\item \href{#method-nodeCentricSteinerForestProblem-getBootstrapSolutionPoolGraphs}{\code{nodeCentricSteinerForestProblem$getBootstrapSolutionPoolGraphs()}}
\item \href{#method-nodeCentricSteinerForestProblem-getSolutionPoolGraphs}{\code{nodeCentricSteinerForestProblem$getSolutionPoolGraphs()}}
\item \href{#method-nodeCentricSteinerForestProblem-findSingleSteinerSolution}{\code{nodeCentricSteinerForestProblem$findSingleSteinerSolution()}}
\item \href{#method-nodeCentricSteinerForestProblem-getInitialSeed}{\code{nodeCentricSteinerForestProblem$getInitialSeed()}}
\item \href{#method-nodeCentricSteinerForestProblem-getLatestSeed}{\code{nodeCentricSteinerForestProblem$getLatestSeed()}}
\item \href{#method-nodeCentricSteinerForestProblem-getAllSeeds}{\code{nodeCentricSteinerForestProblem$getAllSeeds()}}
\item \href{#method-nodeCentricSteinerForestProblem-clone}{\code{nodeCentricSteinerForestProblem$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getConnectivityConstraints"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getConnectivityConstraints'><code>stoneTrees::nodeCentricSteinerTreeProblem$getConnectivityConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getCurrentSolutionGraph"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getCurrentSolutionGraph'><code>stoneTrees::nodeCentricSteinerTreeProblem$getCurrentSolutionGraph()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getCurrentSolutionScore"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getCurrentSolutionScore'><code>stoneTrees::nodeCentricSteinerTreeProblem$getCurrentSolutionScore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getEdgeDT"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getEdgeDT'><code>stoneTrees::nodeCentricSteinerTreeProblem$getEdgeDT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getFixedTerminalConstraints"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getFixedTerminalConstraints'><code>stoneTrees::nodeCentricSteinerTreeProblem$getFixedTerminalConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getNodeDT"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getNodeDT'><code>stoneTrees::nodeCentricSteinerTreeProblem$getNodeDT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getNodeDegreeConstraints"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getNodeDegreeConstraints'><code>stoneTrees::nodeCentricSteinerTreeProblem$getNodeDegreeConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getTerminals"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getTerminals'><code>stoneTrees::nodeCentricSteinerTreeProblem$getTerminals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getTwoCycleConstraints"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getTwoCycleConstraints'><code>stoneTrees::nodeCentricSteinerTreeProblem$getTwoCycleConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getnConnectivityConstraintsCalls"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getnConnectivityConstraintsCalls'><code>stoneTrees::nodeCentricSteinerTreeProblem$getnConnectivityConstraintsCalls()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="isSolutionConnected"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-isSolutionConnected'><code>stoneTrees::nodeCentricSteinerTreeProblem$isSolutionConnected()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="subOptimalSteinerProblem" data-id="getNconnectivityConstraintsCalls"><a href='../../stoneTrees/html/subOptimalSteinerProblem.html#method-subOptimalSteinerProblem-getNconnectivityConstraintsCalls'><code>stoneTrees::subOptimalSteinerProblem$getNconnectivityConstraintsCalls()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="subOptimalSteinerProblem" data-id="getNoveltyConstraints"><a href='../../stoneTrees/html/subOptimalSteinerProblem.html#method-subOptimalSteinerProblem-getNoveltyConstraints'><code>stoneTrees::subOptimalSteinerProblem$getNoveltyConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="subOptimalSteinerProblem" data-id="getOptimumScore"><a href='../../stoneTrees/html/subOptimalSteinerProblem.html#method-subOptimalSteinerProblem-getOptimumScore'><code>stoneTrees::subOptimalSteinerProblem$getOptimumScore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="subOptimalSteinerProblem" data-id="getSolutionPoolScores"><a href='../../stoneTrees/html/subOptimalSteinerProblem.html#method-subOptimalSteinerProblem-getSolutionPoolScores'><code>stoneTrees::subOptimalSteinerProblem$getSolutionPoolScores()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="subOptimalSteinerProblem" data-id="getSolutionTolerance"><a href='../../stoneTrees/html/subOptimalSteinerProblem.html#method-subOptimalSteinerProblem-getSolutionTolerance'><code>stoneTrees::subOptimalSteinerProblem$getSolutionTolerance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="subOptimalSteinerProblem" data-id="identifyMultipleSteinerSolutions"><a href='../../stoneTrees/html/subOptimalSteinerProblem.html#method-subOptimalSteinerProblem-identifyMultipleSteinerSolutions'><code>stoneTrees::subOptimalSteinerProblem$identifyMultipleSteinerSolutions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="subOptimalSteinerProblem" data-id="setSolutionTolerance"><a href='../../stoneTrees/html/subOptimalSteinerProblem.html#method-subOptimalSteinerProblem-setSolutionTolerance'><code>stoneTrees::subOptimalSteinerProblem$setSolutionTolerance()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$new(
  network,
  solverChoice = chooseSolver(),
  verbose = TRUE,
  solverTimeLimit = 300,
  solutionTolerance = 0,
  solverTrace = as.integer(verbose),
  RNGseed = sample.int(size = 1, .Machine$integer.max)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-sampleMultipleBootstrapSteinerSolutions"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-sampleMultipleBootstrapSteinerSolutions}{}}}
\subsection{Method \code{sampleMultipleBootstrapSteinerSolutions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$sampleMultipleBootstrapSteinerSolutions(
  nBootstraps = 5,
  maxItr = 0,
  resamplingProbability = 0.5
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-getBootstrapSolutionPool"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-getBootstrapSolutionPool}{}}}
\subsection{Method \code{getBootstrapSolutionPool()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$getBootstrapSolutionPool()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-getNconnectivityConstraintsCallsPool"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-getNconnectivityConstraintsCallsPool}{}}}
\subsection{Method \code{getNconnectivityConstraintsCallsPool()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$getNconnectivityConstraintsCallsPool()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-getSolutionPool"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-getSolutionPool}{}}}
\subsection{Method \code{getSolutionPool()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$getSolutionPool()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-getBootstrapSolutionPoolGraphs"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-getBootstrapSolutionPoolGraphs}{}}}
\subsection{Method \code{getBootstrapSolutionPoolGraphs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$getBootstrapSolutionPoolGraphs(
  collapseSols = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-getSolutionPoolGraphs"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-getSolutionPoolGraphs}{}}}
\subsection{Method \code{getSolutionPoolGraphs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$getSolutionPoolGraphs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-findSingleSteinerSolution"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-findSingleSteinerSolution}{}}}
\subsection{Method \code{findSingleSteinerSolution()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$findSingleSteinerSolution(maxItr = 20)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-getInitialSeed"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-getInitialSeed}{}}}
\subsection{Method \code{getInitialSeed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$getInitialSeed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-getLatestSeed"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-getLatestSeed}{}}}
\subsection{Method \code{getLatestSeed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$getLatestSeed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-getAllSeeds"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-getAllSeeds}{}}}
\subsection{Method \code{getAllSeeds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$getAllSeeds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerForestProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerForestProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerForestProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
