% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeCentricSteinerTreeProblem-class.R
\name{nodeCentricSteinerTreeProblem}
\alias{nodeCentricSteinerTreeProblem}
\title{Construct an object representation of a Steiner tree/maximum weight connected subgraph (MWCS) problem, with methods to find solutions}
\description{
Solve Steiner problems (MStTP or MWCS) with uniform or no edge weights.
}
\details{
The base stoneTrees class for solving Steiner Tree problems. Each object constructed represents a single Steiner problem
and the associated methods allow for modifications, solution generation etc. See *examples* below.

Input networks must be single component igraph objects with node attributes detailing forced node inclusion in solution ($isTerminal = TRUE)
and/or node costs or prizes for inclusion ($nodeScore).
}
\examples{
library(igraph)

 # Minimum Steiner tree problem

 ## Boolean flags on vertices mark out 'seeds' (nodes that *must* be included)
 unique(V(karateGraph)$isTerminal)
 V(karateGraph)[isTerminal]

 karateMStTP = nodeCentricSteinerTreeProblem$new(karateGraph)
 karateMStTP$findSingleSteinerSolution()

 # Maximum-Weight Connected Subgraph (MWCS)

 ## Vertex attribute details node costs/prizes
 head(V(lymphomaGraph)$nodeScore)

 lymphomaMWCS = nodeCentricSteinerTreeProblem$new(lymphomaGraph)
 lymphomaMWCS$findSingleSteinerSolution()

 # A blend of the two approaches

 ## Say there are some nodes (e.g. drug binding targets) that *must* be included,

 # but otherwise node inclusion should follow node scores
 V(lymphomaGraph)$isTerminal = FALSE
 ## Choose some random nodes to be seeds
 V(lymphomaGraph)[name \%in\% c("4","8","15","16","23","42")]$isTerminal = TRUE

 hybridSteinProb = nodeCentricSteinerTreeProblem$new(lymphomaGraph)
 hybridSteinProb$findSingleSteinerSolution()
}
\references{
Fischetti M, Leitner M, Ljubić I, Luipersbeck M, Monaci M, Resch M, et al. Thinning out Steiner trees: a node-based model for uniform edge costs. Math Program Comput. dimacs11.cs.princeton.edu; 2017;9: 203–229.

Beisser D, Klau GW, Dandekar T, Müller T, Dittrich MT. BioNet: An R-Package for the functional analysis of biological networks. Bioinformatics. 2010;26: 1129–1130.

\url{https://en.wikipedia.org/wiki/Steiner_tree_problem}
}
\seealso{
Other SteinerProblemSolver: 
\code{\link{nodeCentricSteinerForestProblem}},
\code{\link{subOptimalSteinerProblem}}
}
\concept{SteinerProblemSolver}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-nodeCentricSteinerTreeProblem-new}{\code{nodeCentricSteinerTreeProblem$new()}}
\item \href{#method-nodeCentricSteinerTreeProblem-isSolutionConnected}{\code{nodeCentricSteinerTreeProblem$isSolutionConnected()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getCurrentSolutionGraph}{\code{nodeCentricSteinerTreeProblem$getCurrentSolutionGraph()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getCurrentSolutionScore}{\code{nodeCentricSteinerTreeProblem$getCurrentSolutionScore()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getTerminals}{\code{nodeCentricSteinerTreeProblem$getTerminals()}}
\item \href{#method-nodeCentricSteinerTreeProblem-findSingleSteinerSolution}{\code{nodeCentricSteinerTreeProblem$findSingleSteinerSolution()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getNodeDT}{\code{nodeCentricSteinerTreeProblem$getNodeDT()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getEdgeDT}{\code{nodeCentricSteinerTreeProblem$getEdgeDT()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getnConnectivityConstraintsCalls}{\code{nodeCentricSteinerTreeProblem$getnConnectivityConstraintsCalls()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getFixedTerminalConstraints}{\code{nodeCentricSteinerTreeProblem$getFixedTerminalConstraints()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getNodeDegreeConstraints}{\code{nodeCentricSteinerTreeProblem$getNodeDegreeConstraints()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getTwoCycleConstraints}{\code{nodeCentricSteinerTreeProblem$getTwoCycleConstraints()}}
\item \href{#method-nodeCentricSteinerTreeProblem-getConnectivityConstraints}{\code{nodeCentricSteinerTreeProblem$getConnectivityConstraints()}}
\item \href{#method-nodeCentricSteinerTreeProblem-clone}{\code{nodeCentricSteinerTreeProblem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-new}{}}}
\subsection{Method \code{new()}}{
Define a new Steiner Tree problem for solving
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$new(
  network,
  solverChoice = chooseSolver(),
  verbose = TRUE,
  presolveGraph = TRUE,
  solverTimeLimit = 300,
  solverTrace = as.integer(verbose)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{network}}{Search network, with either boolean isTerminal and/or continuous nodeScores recorded for each node in the search network}

\item{\code{solverChoice}}{(optional) Select your preffered solver, or rely on default}

\item{\code{verbose}}{Controls print verbosity of routine}

\item{\code{presolveGraph}}{Whether to include the speed optimisation routine to coalesce adjacent nodes, decreasing the search space (default:TRUE - strongly recommended)}

\item{\code{solverTimeLimit}}{Constrain how long, in seconds, each invocation of the MILP solver can take}

\item{\code{solverTrace}}{Control how much detail to request from the solver}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `nodeCentricSteinerTreeProblem` object, ready to collect solutions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-isSolutionConnected"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-isSolutionConnected}{}}}
\subsection{Method \code{isSolutionConnected()}}{
Query if current solution is connected
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$isSolutionConnected()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
boolean detailing if all nodes in solutionnodes are weakly connected. Empty graphs are disconnected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getCurrentSolutionGraph"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getCurrentSolutionGraph}{}}}
\subsection{Method \code{getCurrentSolutionGraph()}}{
Allow the user to inspect the current solution in the presolved graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getCurrentSolutionGraph()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
igraph induce subgraph of current solution nodes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getCurrentSolutionScore"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getCurrentSolutionScore}{}}}
\subsection{Method \code{getCurrentSolutionScore()}}{
Compute the objective value of the current solution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getCurrentSolutionScore()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A single numeric value - the sum of the nodeScores in proposed solution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getTerminals"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getTerminals}{}}}
\subsection{Method \code{getTerminals()}}{
Detail terminal nodes listed by algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getTerminals()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of terminals; definitions taken from Fischetti et al (2017)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-findSingleSteinerSolution"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-findSingleSteinerSolution}{}}}
\subsection{Method \code{findSingleSteinerSolution()}}{
Initiate a search for a connected solution to the CURRENT constraints. For derived classes this can mean that the solution changes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$findSingleSteinerSolution(maxItr = 20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxItr}}{The maximum number of constraint/solve cycles that the process should attempt}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An induced subgraph of the best solution found given the constraints and interation cap
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getNodeDT"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getNodeDT}{}}}
\subsection{Method \code{getNodeDT()}}{
Provide a table of node details
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getNodeDT()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A table of nodes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getEdgeDT"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getEdgeDT}{}}}
\subsection{Method \code{getEdgeDT()}}{
Provide a table of edge details
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getEdgeDT()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A table of edges in search graph
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getnConnectivityConstraintsCalls"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getnConnectivityConstraintsCalls}{}}}
\subsection{Method \code{getnConnectivityConstraintsCalls()}}{
Provide a list of all connectivity constraint calls
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getnConnectivityConstraintsCalls()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of constraints
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getFixedTerminalConstraints"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getFixedTerminalConstraints}{}}}
\subsection{Method \code{getFixedTerminalConstraints()}}{
Provide a list of fixed terminal constrains
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getFixedTerminalConstraints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of constraints
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getNodeDegreeConstraints"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getNodeDegreeConstraints}{}}}
\subsection{Method \code{getNodeDegreeConstraints()}}{
Provide a list of constraints due to node degrees
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getNodeDegreeConstraints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of constraints
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getTwoCycleConstraints"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getTwoCycleConstraints}{}}}
\subsection{Method \code{getTwoCycleConstraints()}}{
Provide a list of constraints caused by 'two-cycle' constraint - see paper
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getTwoCycleConstraints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of constraints
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-getConnectivityConstraints"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-getConnectivityConstraints}{}}}
\subsection{Method \code{getConnectivityConstraints()}}{
Provide a list of connectivity constraints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$getConnectivityConstraints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of constraints
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeCentricSteinerTreeProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-nodeCentricSteinerTreeProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeCentricSteinerTreeProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
