% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steinLibParser.R
\name{parseSTPfile}
\alias{parseSTPfile}
\title{extract out the edge/arc and terminal information from a STP file}
\usage{
parseSTPfile(fileName)
}
\arguments{
\item{fileName}{The file name of the Steiner problem in SteinLib format}
}
\value{
list A list of two components: 'terminalNodes', which details node terminal information and 'edges', which
}
\description{
A function used to read in SteinLib formatted files. It is not the intention of the author for general users to call this directly.
}
\references{
\url{http://steinlib.zib.de/format.php}
}
