% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steinLibParser.R
\name{readMStTPgraph}
\alias{readMStTPgraph}
\title{create an igraph representation of a MStT problem}
\usage{
readMStTPgraph(MStTPstpFile)
}
\arguments{
\item{MStTPstpFile}{The URI of the Steiner problem in SteinLib format}
}
\value{
steinerProblem An igraph object with vertex names and a per vertex flag 'isTerminal' which specifies if a node is a terminal or not.
  Note that upon creation of the igraph object, node names (as defined in the STP file) and nodeIDs (as produced by igraph).
}
\description{
The SteinLib resource provides a reference collection of Steiner Tree problems and solutions. Very useful for validation. This
function allows you to read in graphs relating to the minimum Steiner Tree problem.
}
\references{
\url{http://steinlib.zib.de/format.php}
}
