% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steinLibParser.R
\name{readMWCSgraph}
\alias{readMWCSgraph}
\title{create an igraph representation of a MWCS problem}
\usage{
readMWCSgraph(MWCSstpFile)
}
\arguments{
\item{MWCSstpFile}{The URI of the Steiner problem in SteinLib format}
}
\value{
steinerProblem An igraph object with vertex names and a per vertex flag 'isTerminal' which specifies if a node is a terminal or not.
  Note that upon creation of the igraph object, node names (as defined in the STP file) and nodeIDs (as produced by igraph).
}
\description{
The SteinLib resource provides a reference collection of Steiner Tree problems and solutions. Very useful for validation. This function allows
you to read in graphs relating to maximum-weight connected subgraph problems.
}
\references{
\url{http://steinlib.zib.de/format.php}
}
