% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoneTrees-package.R
\docType{package}
\name{stoneTrees-package}
\alias{stoneTrees-package}
\title{stoneTrees: Solve Node Centric Steiner Tree Problems}
\description{
A package aimed at solving Steiner ('stone') tree problems with uniform edge weights, such as the Minimum Steiner Tree (MStT) and Maximum-Weight Connected Subgraph (MWCS) problems. It serves as an R implementation of the paper Fischetti M et al., Thinning out Steiner trees: a node-based model for uniform edge costs. Math Program Comp (2017).
}
\author{
\strong{Maintainer}: Adam Sardar \email{adam@sardar.me.uk} (\href{https://orcid.org/0000-0002-6912-6125}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Julie Borgel \email{borgel.julie@gmail.com} [contributor]
  \item Ines De Santiago \email{ines.desantiago@etherapeutics.co.uk} (\href{https://orcid.org/0000-0002-8172-1322}{ORCID}) [contributor]
}

}
\keyword{internal}
