% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toCPXMatrix.R
\name{toCPXMatrix}
\alias{toCPXMatrix}
\title{Function to map a constraint matrix to a cplex compatible row,column, value set of triples}
\usage{
toCPXMatrix(Amat)
}
\arguments{
\item{Amat}{Constraint matrix. Can be of type sparseMatrix or simple_triplet_matrix}
}
\value{
A list compatible with the copyLpCPLEX in cplexAPI
}
\description{
This function is copied wholesale from Rcplex: See Rcplex:::toCPXMatrix. However, Rcplex is a difficult package
to include as a dependency as it requires explicit compilation with links to proprietry header/library files.
}
\seealso{
cplexAPI::copyLpCPLEX
}
