% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPresolveByCondensation.R
\name{uncondenseGraph}
\alias{uncondenseGraph}
\title{Decondense a graph following a procedure to group together prize nodes. If the input was not condensed, then we
just return it back}
\usage{
uncondenseGraph(condensedGraph)
}
\arguments{
\item{condensedGraph}{An igraph representation of the condensed graph (where prize nodes). The result of calling 'condenseGraph()'}
}
\value{
uncondensed graph
}
\description{
Decondense a graph following a procedure to group together prize nodes. If the input was not condensed, then we
just return it back
}
\seealso{
condenseSearchGraph
}
