% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steinLibParser.R
\name{writeMStTPfile_heinzFormat}
\alias{writeMStTPfile_heinzFormat}
\title{create an igraph representation of a MStT problem}
\usage{
writeMStTPfile_heinzFormat(
  network.igraph,
  file.name,
  remark = "Nothing of note",
  problem = "MStTP",
  name = "Autogenerated Name"
)
}
\arguments{
\item{network.igraph}{Graph object to be written out}

\item{file.name}{The file name of the output file}

\item{remark}{SteinLib format has a remark field (rarely used), but feel free to add something in here}

\item{problem}{SteinLib format has a problem field (rarely used), but feel free to add something in here}

\item{name}{SteinLib format has a name field (rarely used), but feel free to add something in here}
}
\value{
node2ID A data.table of nodeName to nodeIndex mappings
}
\description{
The heinz program uses a slightly modified version of steinLib format, with no documentation. This function
  writes out a MWCS problem into an appropriate form for use with heinz
}
\references{
\url{http://steinlib.zib.de/format.php}

\url{https://github.com/ls-cwi/heinz}
}
