% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPresolveByCondensation.R
\name{condenseSearchGraph}
\alias{condenseSearchGraph}
\title{Collapse neighbouring terminal and potential terminal nodes together}
\usage{
condenseSearchGraph(graphToCondense, condensedNodeSep = ";")
}
\arguments{
\item{graphToCondense}{igraph network that we wish to contract}

\item{condensedNodeSep}{(optional) Specify the value to separate node names with}
}
\value{
A graph with neighbouring terminal and potential terminal nodes collapsed to the same node
}
\description{
Part of the Steiner Tree solving method involves enumerating all vs all node separators of nodes in differing
components. This scales very, very badly and for some runs can actually be the major computational bottleneck. It's
also unnecessary as neighbouring prize nodes and terminals will always be in the resultant graph.
}
\seealso{
uncondenseGraph
}
