% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoneTrees-package.R
\docType{package}
\name{stoneTrees}
\alias{stoneTrees}
\alias{stoneTrees-package}
\title{R implementation of Steiner Tree solver}
\description{
A package dedicated to finding minimum Steiner trees in networks. Particularly biological ones,
which tend to be very sparse and on the easier end of the spectrum. This package is particularly
interested in Minimum Steiner Tree (MStT), Maximum-Weight Connected Subgraph (MWCS) and
Prize-Collecting Steiner Tree (PCST) problems.
}
\details{
This package serves as a faithful implementation of "Thinning out Steiner Trees".
}
\author{
Adam Sardar
}
