% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subOptimalNodeCentricSteinerTreeProblem-class.R
\name{subOptimalSteinerProblem}
\alias{subOptimalSteinerProblem}
\title{Construct an object representation of a multiple-solution Steiner tree/maximum weight connected subgraph (MWCS) problem}
\description{
Collect degenerate and sub-optimal solutions to Steiner problems (MStTP or MWCS) with uniform or no edge weights.
}
\details{
Rather than find just a single solution to a MStTP/MWCS, one can populate a solution pool with multiple degenerate/tolerable solutions.

This class is derived from *nodeCentricSteinerTreeProblem*: all methods available in the superclass are available here. The difference is that after
each acceptable solution is found, the solution is a.) stored in a solution pool and b.) used to generate a 'novelty' constraint on future solutions.
}
\examples{
library(igraph)

 # Maximum-Weight Connected Subgraph (MWCS) - find sub-optimal solutions

 ## Vertex attribute details node costs/prizes
 head(V(lymphomaGraph)$nodeScore)

 lymphoma_multiMWCS = subOptimalSteinerProblem$new(lymphomaGraph, solutionTolerance = 0.5)

 #Populate the solution pool with multiple solutions - notice the
 lymphoma_multiMWCS$identifyMultipleSteinerSolutions()

 lymphoma_multiMWCS$getSolutionPoolGraphs(collapseSols = FALSE)

 lymphoma_multiMWCS$getSolutionPoolScores()

 #All solution scores are within tolerance
 diff(range(lymphoma_multiMWCS$getSolutionPoolScores()))
}
\references{
Fischetti M, Leitner M, Ljubić I, Luipersbeck M, Monaci M, Resch M, et al. Thinning out Steiner trees: a node-based model for uniform edge costs. Math Program Comput. dimacs11.cs.princeton.edu; 2017;9: 203–229.

Beisser D, Klau GW, Dandekar T, Müller T, Dittrich MT. BioNet: An R-Package for the functional analysis of biological networks. Bioinformatics. 2010;26: 1129–1130.

\url{https://en.wikipedia.org/wiki/Steiner_tree_problem}
}
\seealso{
nodeCentricSteinerTreeProblem

Other SteinerProblemSolver: 
\code{\link{nodeCentricSteinerForestProblem}},
\code{\link{nodeCentricSteinerTreeProblem}}
}
\concept{SteinerProblemSolver}
\section{Super class}{
\code{\link[stoneTrees:nodeCentricSteinerTreeProblem]{stoneTrees::nodeCentricSteinerTreeProblem}} -> \code{subOptimalSteinerProblem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-subOptimalSteinerProblem-new}{\code{subOptimalSteinerProblem$new()}}
\item \href{#method-subOptimalSteinerProblem-getSolutionPool}{\code{subOptimalSteinerProblem$getSolutionPool()}}
\item \href{#method-subOptimalSteinerProblem-getSolutionPoolGraphs}{\code{subOptimalSteinerProblem$getSolutionPoolGraphs()}}
\item \href{#method-subOptimalSteinerProblem-getSolutionPoolScores}{\code{subOptimalSteinerProblem$getSolutionPoolScores()}}
\item \href{#method-subOptimalSteinerProblem-getOptimumScore}{\code{subOptimalSteinerProblem$getOptimumScore()}}
\item \href{#method-subOptimalSteinerProblem-getNoveltyConstraints}{\code{subOptimalSteinerProblem$getNoveltyConstraints()}}
\item \href{#method-subOptimalSteinerProblem-getSolutionTolerance}{\code{subOptimalSteinerProblem$getSolutionTolerance()}}
\item \href{#method-subOptimalSteinerProblem-setSolutionTolerance}{\code{subOptimalSteinerProblem$setSolutionTolerance()}}
\item \href{#method-subOptimalSteinerProblem-getNconnectivityConstraintsCalls}{\code{subOptimalSteinerProblem$getNconnectivityConstraintsCalls()}}
\item \href{#method-subOptimalSteinerProblem-identifyMultipleSteinerSolutions}{\code{subOptimalSteinerProblem$identifyMultipleSteinerSolutions()}}
\item \href{#method-subOptimalSteinerProblem-clone}{\code{subOptimalSteinerProblem$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="findSingleSteinerSolution"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-findSingleSteinerSolution'><code>stoneTrees::nodeCentricSteinerTreeProblem$findSingleSteinerSolution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getConnectivityConstraints"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getConnectivityConstraints'><code>stoneTrees::nodeCentricSteinerTreeProblem$getConnectivityConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getCurrentSolutionGraph"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getCurrentSolutionGraph'><code>stoneTrees::nodeCentricSteinerTreeProblem$getCurrentSolutionGraph()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getCurrentSolutionScore"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getCurrentSolutionScore'><code>stoneTrees::nodeCentricSteinerTreeProblem$getCurrentSolutionScore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getEdgeDT"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getEdgeDT'><code>stoneTrees::nodeCentricSteinerTreeProblem$getEdgeDT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getFixedTerminalConstraints"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getFixedTerminalConstraints'><code>stoneTrees::nodeCentricSteinerTreeProblem$getFixedTerminalConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getNodeDT"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getNodeDT'><code>stoneTrees::nodeCentricSteinerTreeProblem$getNodeDT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getNodeDegreeConstraints"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getNodeDegreeConstraints'><code>stoneTrees::nodeCentricSteinerTreeProblem$getNodeDegreeConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getTerminals"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getTerminals'><code>stoneTrees::nodeCentricSteinerTreeProblem$getTerminals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getTwoCycleConstraints"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getTwoCycleConstraints'><code>stoneTrees::nodeCentricSteinerTreeProblem$getTwoCycleConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="getnConnectivityConstraintsCalls"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-getnConnectivityConstraintsCalls'><code>stoneTrees::nodeCentricSteinerTreeProblem$getnConnectivityConstraintsCalls()</code></a></span></li>
<li><span class="pkg-link" data-pkg="stoneTrees" data-topic="nodeCentricSteinerTreeProblem" data-id="isSolutionConnected"><a href='../../stoneTrees/html/nodeCentricSteinerTreeProblem.html#method-nodeCentricSteinerTreeProblem-isSolutionConnected'><code>stoneTrees::nodeCentricSteinerTreeProblem$isSolutionConnected()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the subOptimalSteinerProblem class. Alongside the arguments for the super-class constructor, there is also 'solutionTolerance', which instructs the object as to the gap between optimal and observed solution that is acceptable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$new(
  network,
  solverChoice = chooseSolver(),
  verbose = TRUE,
  presolveGraph = TRUE,
  solverTimeLimit = 300,
  solutionTolerance = 0,
  solverTrace = as.integer(verbose)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{network}}{Search network, with either boolean isTerminal and/or continuous nodeScores recorded for each node in the search network}

\item{\code{solverChoice}}{(optional) Select your preffered solver, or rely on default}

\item{\code{verbose}}{Controls print verbosity of routine}

\item{\code{presolveGraph}}{Whether to include the speed optimisation routine to coalesce adjacent nodes, decreasing the search space (default:TRUE - strongly recommended)}

\item{\code{solverTimeLimit}}{Constrain how long, in seconds, each invocation of the MILP solver can take}

\item{\code{solutionTolerance}}{If sub-optimal solutions are to be collected, what tolerance to use? (default: 0)}

\item{\code{solverTrace}}{Control how much detail to request from the solver}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `nodeCentricSteinerForestProblem` object, ready to collect solutions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-getSolutionPool"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-getSolutionPool}{}}}
\subsection{Method \code{getSolutionPool()}}{
Internal function aggegating solutions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$getSolutionPool()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A set of integers that map to nodes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-getSolutionPoolGraphs"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-getSolutionPoolGraphs}{}}}
\subsection{Method \code{getSolutionPoolGraphs()}}{
Internal function aggegating solutions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$getSolutionPoolGraphs(collapseSols = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collapseSols}}{Return all graphs collapsed, or a list of graphs}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A set of integers that map to nodes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-getSolutionPoolScores"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-getSolutionPoolScores}{}}}
\subsection{Method \code{getSolutionPoolScores()}}{
Compute the scores of the solutions in the solution pool. These are in the same order as the list of graphs returned by $getSolutionPoolGraphs(FALSE)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$getSolutionPoolScores()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A set of continuous scores of solutions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-getOptimumScore"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-getOptimumScore}{}}}
\subsection{Method \code{getOptimumScore()}}{
Returns the optimum score from solutions in the solution pool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$getOptimumScore()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Continuous value of the top score
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-getNoveltyConstraints"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-getNoveltyConstraints}{}}}
\subsection{Method \code{getNoveltyConstraints()}}{
Internal utility function that forces new solutions to be found
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$getNoveltyConstraints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Matrix of existing solutions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-getSolutionTolerance"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-getSolutionTolerance}{}}}
\subsection{Method \code{getSolutionTolerance()}}{
setter for the solver solution tolerance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$getSolutionTolerance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Continuous value of the top score
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-setSolutionTolerance"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-setSolutionTolerance}{}}}
\subsection{Method \code{setSolutionTolerance()}}{
setter for the solver solution tolerance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$setSolutionTolerance(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Single continuous number to use for search}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-getNconnectivityConstraintsCalls"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-getNconnectivityConstraintsCalls}{}}}
\subsection{Method \code{getNconnectivityConstraintsCalls()}}{
Provide a list of all connectivity constraint calls
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$getNconnectivityConstraintsCalls()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of constraints
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-identifyMultipleSteinerSolutions"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-identifyMultipleSteinerSolutions}{}}}
\subsection{Method \code{identifyMultipleSteinerSolutions()}}{
The multi-solution version of `$findSingleSteinerSolution`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$identifyMultipleSteinerSolutions(maxItr = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxItr}}{As we sample the sub-optimal solutions, what is the maximum number of within-tolerance solutions to aggregate?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subOptimalSteinerProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-subOptimalSteinerProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{subOptimalSteinerProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
