% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPresolveByCondensation.R
\name{uncondenseGraph}
\alias{uncondenseGraph}
\title{Decondense a graph following a procedure to group together prize nodes. If the input was not condensed, then we
just return it back}
\usage{
uncondenseGraph(condensedGraph, originalGraph = NULL,
  nodeNameSep = ";")
}
\arguments{
\item{condensedGraph}{An igraph representation of the condensed graph (where prize nodes )}

\item{originalGraph}{(optional) The original graph that the condensation procedure was run on. Should probably be left
empty as the function will evaluate and draw in the original network}

\item{nodeNameSep}{(optional) If a different node-name separator was used in the graph condensation step}
}
\value{
uncondensed graph
}
\description{
Decondense a graph following a procedure to group together prize nodes. If the input was not condensed, then we
just return it back
}
\seealso{
condenseSearchGraph
}
